spilit関数をutilityに移動
